requirements.txt
agate==1.9.1
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.1.0
babel==2.17.0
cattrs==24.1.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
daff==1.3.46
dataclasses==0.6
dbt-adapters==1.14.0
dbt-common==1.15.0
dbt-core==1.9.2
dbt-extractor==0.5.1
dbt-postgres==1.9.0
dbt-semantic-interfaces==0.7.4
debugpy==1.8.12
decorator==5.2.1
deepdiff==7.0.1
dnspython==2.7.0
executing==2.2.0
fonttools==4.56.0
idna==3.10
importlib-metadata==6.11.0
ipykernel==6.29.5
ipython==9.0.1
ipython_pygments_lexers==1.1.1
isodate==0.6.1
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
leather==0.4.0
MarkupSafe==3.0.2
mashumaro==3.14
matplotlib==3.10.0
matplotlib-inline==0.1.7
more-itertools==10.6.0
msgpack==1.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.3
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
pandera==0.23.1
parsedatetime==2.6
parso==0.8.4
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pymongo==4.11.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.1
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
requests==2.32.3
requests-cache==1.2.1
rpds-py==0.22.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
snowplow-tracker==1.0.4
sqlparse==0.5.3
stack-data==0.6.3
text-unidecode==1.3
threadpoolctl==3.6.0
tornado==6.4.2
traitlets==5.14.3
typeguard==4.4.2
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
url-normalize==1.4.3
urllib3==2.3.0
wcwidth==0.2.13
zipp==3.21.0